requirements.txt
opencv-python>=4.8.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
RPi.GPIO>=0.7.1
